+++ src/hamster/storage/storage.py
@@ -213,6 +213,9 @@ class Storage(object):
     def get_ext_activities(self, search = ""):
         return self.__get_ext_activities(search)
 
+    def export_fact(self, fact_id):
+        return self.__export_fact(fact_id)
+
     def change_category(self, id, category_id):
         changed = self.__change_category(id, category_id)
         if changed: